Version 1.3 by Stephen Martin.
https://ftp.gwdg.de/pub/x11/x.org/contrib/applications/sunclock1.3.tar.Z

Version 1.3, November 16, 1994.

- Timezone now displayed in icon. Suggested by Rich Gonzalez
    (gonzo@u.washington.edu)
- Should now work under Linux/XFree86 - Wolfram Gloger
    (wmglo@Dent.MED.Uni-Muenchen.DE)

Version 1.2, October 13, 1994.

- Used new icon suppied by Anthony Thyssen (anthony@cit.gu.edu.au). I can't
  see a difference can you see a difference.
- Fixed stuff for HP using NEW_CTIME ifdef as suggested by Richard Lloyd
  (R.K.Lloyd@csc.liv.ac.uk). Untested by me on HP.
- Support for Solaris thanks to Dave Yearke (yearke@calspan.com) who also
  pointed out some other bugs and fixes.

==============================================================================

Version 1.1, August 31, 1994.

This is version 1.1 of sunclock. New in this version is support for an
interactive map. This allows you to define places on the map in your
~/.sunclockrc file. These places will show up labelled in the large map,
(not the icon). Clicking on a place will change the clock readout to the
timezone of the that place. See the man page for details.

Stephen Martin, smartin@fujitsu.ca - August 31, 1994.